renamed file, removed bad code, adjust cal values
